db: get rid of code to replace session object
This was causing a new session object to be created anytime a new app was initialized which would leave existing sessions and transactions open with no discernible way to close them (until GC runs I believe).

I'm not sure why this code was here in the first place but it doesn't seem to add any value.  If you want split scope sessions, you can easily do that with Flask-WebTest.